+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -416,13 +416,23 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 4000
 		 * @public
 		 */
-		titleHideDelay: PropTypes.number
+		titleHideDelay: PropTypes.number,
+
+		/**
+		 * The amount of time in miliseconds that should pass before the tooltip disappears from the
+		 * controls. Setting this to `0` disables the hiding.
+		 *
+		 * @type {Number}
+		 * @default 3000
+		 * @public
+		 */
+		tooltipHideDelay: PropTypes.number
 	}
 
 	static defaultProps = {
 		autoCloseTimeout: 7000,
 		backwardIcon: 'backward',
-		feedbackHideDelay: 2000,
+		feedbackHideDelay: 3000,
 		forwardIcon: 'forward',
 		jumpBackwardIcon: 'skipbackward',
 		jumpBy: 30,
@@ -440,7 +450,8 @@ const VideoPlayerBase = class extends React.Component {
 			slowRewind: ['-1/2', '-1']
 		},
 		playIcon: 'play',
-		titleHideDelay: 4000
+		titleHideDelay: 4000,
+		tooltipHideDelay: 3000
 	}
 
 	constructor (props) {
@@ -494,7 +505,8 @@ const VideoPlayerBase = class extends React.Component {
 			proportionPlayed: 0,
 			sliderScrubbing: false,
 			sliderKnobProportion: 0,
-			titleVisible: true
+			titleVisible: true,
+			tooltipVisible: true
 		};
 	}
 
@@ -632,17 +644,23 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	showControls = () => {
+		this.startDelayedFeedbackHide();
 		this.startDelayedTitleHide();
+		this.startDelayedTooltipHide();
 		forwardControlsAvailable({available: true}, this.props);
 		this.setState({
 			bottomControlsRendered: true,
 			bottomControlsVisible: true,
-			titleVisible: true
+			feedbackVisible: true,
+			titleVisible: true,
+			tooltipVisible: true
 		});
 	}
 
 	hideControls = () => {
+		this.stopDelayedFeedbackHide();
 		this.stopDelayedTitleHide();
+		this.stopDelayedTooltipHide();
 		forwardControlsAvailable({available: false}, this.props);
 		this.setState({bottomControlsVisible: false, more: false});
 	}
@@ -665,6 +683,22 @@ const VideoPlayerBase = class extends React.Component {
 
 	hideTitleJob = new Job(this.hideTitle)
 
+	startDelayedTooltipHide = () => {
+		if (this.props.tooltipHideDelay) {
+			this.hideTooltipJob.startAfter(this.props.tooltipHideDelay);
+		}
+	}
+
+	stopDelayedTooltipHide = () => {
+		this.hideTooltipJob.stop();
+	}
+
+	hideTooltip = () => {
+		this.setState({tooltipVisible: false});
+	}
+
+	hideTooltipJob = new Job(this.hideTooltip)
+
 	startDelayedFeedbackHide = () => {
 		if (this.props.feedbackHideDelay) {
 			this.hideFeedbackJob.startAfter(this.props.feedbackHideDelay);
@@ -684,14 +718,6 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	hideFeedbackJob = new Job(this.hideFeedback)
-
-	showFeedback = () => {
-		this.setState({feedbackVisible: true});
-	}
-
-	hideFeedback = () => {
-		this.setState({feedbackVisible: false});
-	}
 
 	//
 	// Media Interaction Methods
@@ -749,8 +775,12 @@ const VideoPlayerBase = class extends React.Component {
 	 * @private
 	 */
 	send = (action, props) => {
-		this.showFeedback();
+		this.setState({
+			feedbackVisible: true,
+			tooltipVisible: true
+		});
 		this.startDelayedFeedbackHide();
+		this.startDelayedTooltipHide();
 		this.video[action](props);
 	}
 
@@ -1071,6 +1101,15 @@ const VideoPlayerBase = class extends React.Component {
 		this.sliderScrubbing = ev.detached;
 		this.sliderKnobProportion = ev.proportion;
 	}
+	handleMouseMove = () => this.showTooltipJob.throttle()
+	showTooltipJob = new Job(() => {
+		this.setState({
+			feedbackVisible: false,
+			tooltipVisible: true
+		});
+		this.stopDelayedFeedbackHide();
+		this.startDelayedTooltipHide();
+	}, 50)
 	onJumpBackward = (ev) => {
 		ev = this.addStateToEvent(ev);
 		forwardJumpBackwardButtonClick(ev, this.props);
@@ -1143,6 +1182,7 @@ const VideoPlayerBase = class extends React.Component {
 		delete rest.onPlayButtonClick;
 		delete rest.playbackRateHash;
 		delete rest.titleHideDelay;
+		delete rest.tooltipHideDelay;
 
 		// Remove the events we manually added so they aren't added twice or fail.
 		for (let eventName in handledCustomMediaEventsMap) {
@@ -1190,10 +1230,11 @@ const VideoPlayerBase = class extends React.Component {
 								value={this.state.proportionPlayed}
 								onChange={this.onSliderChange}
 								onKnobMove={this.handleKnobMove}
+								onMouseMove={this.handleMouseMove}
 								onSpotlightUp={this.hideControls}
 								onSpotlightDown={this.handleSpotlightDownFromSlider}
 							>
-								<div className={css.sliderTooltip}>
+								<div className={css.sliderTooltip} style={{display: this.state.tooltipVisible ? 'block' : 'none'}}>
 									<Feedback playbackState={this.prevCommand} visible={this.state.feedbackVisible} >
 										{this.selectPlaybackRate(this.speedIndex)}
 									</Feedback>